+++ test/test_extensions.rb
@@ -14,6 +14,10 @@ describe Sidekiq::Extensions do
     def self.long_class_method
       raise "Should not be called!"
     end
+
+    def self.long_class_method_with_optional_args(*arg, **kwargs)
+      raise "Should not be called!"
+    end
   end
 
   it 'allows delayed execution of ActiveRecord class methods' do
@@ -21,6 +25,15 @@ describe Sidekiq::Extensions do
     q = Sidekiq::Queue.new
     assert_equal 0, q.size
     MyModel.delay.long_class_method
+    assert_equal ['default'], Sidekiq::Queue.all.map(&:name)
+    assert_equal 1, q.size
+  end
+
+  it 'allows delayed execution of ActiveRecord class methods with optional arguments' do
+    assert_equal [], Sidekiq::Queue.all.map(&:name)
+    q = Sidekiq::Queue.new
+    assert_equal 0, q.size
+    MyModel.delay.long_class_method_with_optional_args(with: :keywords)
     assert_equal ['default'], Sidekiq::Queue.all.map(&:name)
     assert_equal 1, q.size
   end
@@ -53,6 +66,10 @@ describe Sidekiq::Extensions do
     def greetings(a, b)
       raise "Should not be called!"
     end
+
+    def greetings_with_optional_args(*arg, **kwargs)
+      raise "Should not be called!"
+    end
   end
 
   it 'allows delayed delivery of ActionMailer mails' do
@@ -60,6 +77,15 @@ describe Sidekiq::Extensions do
     q = Sidekiq::Queue.new
     assert_equal 0, q.size
     UserMailer.delay.greetings(1, 2)
+    assert_equal ['default'], Sidekiq::Queue.all.map(&:name)
+    assert_equal 1, q.size
+  end
+
+  it 'allows delayed delivery of ActionMailer mails with optional arguments' do
+    assert_equal [], Sidekiq::Queue.all.map(&:name)
+    q = Sidekiq::Queue.new
+    assert_equal 0, q.size
+    UserMailer.delay.greetings_with_optional_args(with: :keywords)
     assert_equal ['default'], Sidekiq::Queue.all.map(&:name)
     assert_equal 1, q.size
   end
@@ -81,6 +107,9 @@ describe Sidekiq::Extensions do
   class SomeClass
     def self.doit(arg)
     end
+
+    def self.doit_with_optional_args(*arg, **kwargs)
+    end
   end
 
   it 'allows delay of any ole class method' do
@@ -90,8 +119,18 @@ describe Sidekiq::Extensions do
     assert_equal 1, q.size
   end
 
+  it 'allows delay of any ole class method with optional arguments' do
+    q = Sidekiq::Queue.new
+    assert_equal 0, q.size
+    SomeClass.delay.doit_with_optional_args(with: :keywords)
+    assert_equal 1, q.size
+  end
+
   module SomeModule
     def self.doit(arg)
+    end
+
+    def self.doit_with_optional_args(*arg, **kwargs)
     end
   end
 
@@ -109,4 +148,10 @@ describe Sidekiq::Extensions do
     assert_equal 1, q.size
   end
 
+  it 'allows delay of any module class method with optional arguments' do
+    q = Sidekiq::Queue.new
+    assert_equal 0, q.size
+    SomeModule.delay.doit_with_optional_args(with: :keywords)
+    assert_equal 1, q.size
+  end
 end